+++ tests/e2e/helpers.ts
@@ -154,55 +154,3 @@ export function useHasPermission({
     return isAdmin || permissions.includes(permission);
   };
 }
-
-interface ClientCreateParams {
-  page: Page;
-  clientName?: string;
-  assignTo?: string;
-  withNavigation?: boolean;
-  createIfNotExist?: boolean;
-}
-export const createClient = async (params: ClientCreateParams) => {
-  const {
-    page,
-    clientName,
-    assignTo,
-    withNavigation = true,
-    createIfNotExist = false,
-  } = params;
-
-  if (withNavigation) {
-    await page
-      .locator('[data-cy="navigationBar"]')
-      .getByRole('link', { name: 'Clients', exact: true })
-      .click();
-  }
-
-  const doRecordsExist = await page.getByText('No records found').isHidden();
-
-  if (createIfNotExist && doRecordsExist) {
-    return;
-  }
-
-  await page
-    .getByRole('main')
-    .getByRole('link', { name: 'New Client' })
-    .click();
-
-  await page.locator('#name').fill(clientName || 'Company Name');
-  await page.locator('#first_name_0').fill('First Name');
-  await page.locator('#last_name_0').fill('Last Name');
-  await page.locator('#email_0').fill('first@example.com');
-
-  if (assignTo) {
-    await page
-      .locator('select[id="assigned_user_id"]')
-      .selectOption({ label: assignTo });
-  }
-
-  await page.getByRole('button', { name: 'Save' }).click();
-
-  await expect(
-    page.getByText('Successfully created client', { exact: true })
-  ).toBeVisible();
-};
